LPS-125879 Export a new util to construct the new DataRecordValues from fields
