+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-renderer/src/main/resources/META-INF/resources/index.js
@@ -24,6 +24,7 @@ export {
 export {PagesVisitor, RulesVisitor} from './js/util/visitors.es';
 export * as FormSupport from './js/util/FormSupport.es';
 export {getConnectedReactComponentAdapter} from './js/util/ReactComponentAdapter.es';
+export {default as setDataRecord} from './js/util/setDataRecord.es';
 
 // Composing Form
